+++ src/tests/test_core.py
@@ -1,7 +1,7 @@
 import unittest
 
 from figgler.core import FIG_ENV_REGEX, figvars, FigURI, container_name, groupby_containers, \
-    container_ordinal, Containers
+    split_uri, squash_ports, container_ordinal, Containers
 
 
 class FigenvRegexTestCase(unittest.TestCase):
@@ -38,41 +38,54 @@ class FigvarsTestCase(unittest.TestCase):
         self.assertDictEqual(figvars(allvars), expected)
 
 
-class FigURITestCase(unittest.TestCase):
+class splitUriTestCase(unittest.TestCase):
 
     def testWillSplitUrlIntoProperChunks(self):
         uri = "tcp://172.17.0.8:5000"
         expected = {
-            'uri': 'tcp://172.17.0.8:5000',
             'protocol': 'tcp',
             'host': '172.17.0.8',
             'port': '5000'
         }
-        f = FigURI(uri)
-        actual = {'uri': f.uri, 'protocol': f.protocol, 'host': f.host, 'port': f.port}
-        self.assertDictEqual(actual, expected)
+        self.assertDictEqual(split_uri(uri), expected)
 
     def testWillRaiseValueErrorIfGivenUnprocessibleURI(self):
         uri = "thisis:/:some::bollocks"
         with self.assertRaises(ValueError):
-            FigURI(uri)
+            split_uri(uri)
+
+
+class squashPortsTestCase(unittest.TestCase):
+
+    def testWillSquashIntoOneDictsBasingOnPortKey(self):
+        containers = [
+            {'port': 90, 'host': '172.2.0.1', 'protocol': 'tcp'},
+            {'port': 91, 'host': '172.2.0.1', 'protocol': 'tcp'}
+        ]
+        expected = [{'ports': [90, 91], 'host': '172.2.0.1', 'protocol': 'tcp'}]
+        self.assertListEqual(expected[0]['ports'], squash_ports(containers)[0]['ports'])
+
+
+class FigURITestCase(unittest.TestCase):
+
+    pass
 
 
 class ContainerNameTestCase(unittest.TestCase):
 
     def testWillSuccessfulyExtractContainerNameFromSingleUnderscoreName(self):
-        key = 'DB_12_PORT'
+        key = 'DB_12_PORT_5000_TCP'
         self.assertEquals('db', container_name(key))
 
     def testWillSuccessfulyExtractContainerNamesFromMultiUnderscoreName(self):
-        key = 'SOME_OTHER_SERVICE_1_PORT'
+        key = 'SOME_OTHER_SERVICE_1_PORT_3000_TCP'
         self.assertEquals('some_other_service', container_name(key))
 
 
 class ContainerOrdinalTestCase(unittest.TestCase):
 
     def testWillReturnOrdinal(self):
-        key = 'SOME_SERVICE_13_PORT'
+        key = 'SOME_SERVICE_13_PORT_3000_TCP'
         self.assertEquals(13, container_ordinal(key))
 
 
@@ -80,31 +93,36 @@ class GroupbyContainersTestCase(unittest.TestCase):
 
     def testGivenDictKeyedWithFigEnvvarsWillGroupValuesByContainerName(self):
         environ = {
-            'DB_1_PORT': "whatever",
-            'DB_2_PORT': "whatever",
-            'REDIS_3_PORT': "redis...",
+            'DB_1_PORT_80_TCP': 'tcp://172.2.0.1:5000',
+            'DB_2_PORT_90_TCP': 'tcp://172.2.0.1:5001',
+            'REDIS_3_PORT_30_TCP': 'tcp://172.2.0.1:5002',
         }
         self.assertSetEqual(set(groupby_containers(environ).keys()), {'db', 'redis'})
 
-    def testGroupsWillBeSortedBasingOnKeyNumber(self):
+    def testPortsWillBeMergedIntoListsAndSortedAscending(self):
         environ = {
-            'DB_1_PORT': "firstdb",
-            'DB_2_PORT': "seconddb",
-            'REDIS_3_PORT': "secondredis",
-            'REDIS_2_PORT': "firstredis",
+            'DB_1_PORT_5000_TCP': "tcp://172.0.2.1:5000",
+            'DB_1_PORT_6000_TCP': "tcp://172.0.2.1:6000",
+            'HTTP_1_PORT_80_TCP': "tcp://172.0.2.3:80",
+            'HTTP_2_PORT_80_TCP': "tcp://172.0.2.4:80",
         }
         grouped = groupby_containers(environ)
-        self.assertListEqual(grouped['db'], ["firstdb", "seconddb"])
-        self.assertListEqual(grouped['redis'], ["firstredis", "secondredis"])
+        self.assertListEqual(grouped['db'][0]['ports'], ['5000', '6000'])
+        self.assertListEqual(grouped['http'][0]['ports'], ['80'])
+        self.assertListEqual(grouped['http'][1]['ports'], ['80'])
 
 
 class ContainersTestCase(unittest.TestCase):
 
     def testWillMapFigURIOverGivenKwargsAndExtendSelfWithThen(self):
         kwargs = {
-            'db': ['tcp://172.12.0.1:5432'],
-            'memcached': ['tcp://172.12.0.2:3000', 'tcp://172.12.0.3:3001']
+            'db': [
+                {'protocol': 'tcp', 'host': '172.12.0.1', 'ports': ['5432']}
+            ],
+            'memcached': [
+                {'protocol': 'tcp', 'host': '172.12.0.2', 'ports': ['3001', '3000']},
+                {'protocol': 'tcp', 'host': '172.12.0.3', 'ports': ['3001', '3000']}
+            ]
         }
         containers = Containers(**kwargs)
-        self.assertEqual(containers.memcached[0].uri, 'tcp://172.12.0.2:3000')
-        self.assertEqual(containers.memcached[0].port, '3000')
+        self.assertListEqual(containers.memcached[0].ports, ['3000', '3001'])